removed stack trace added existence check

+++ src/main/java/com/alinorouzi/cardemo/service/impl/CarService.java
@@ -52,6 +52,8 @@ public class CarService implements ICarService {
 
     @Override
     public void delete(Integer id) {
-        carRepository.deleteById(id);
+        if (carRepository.existsById(id)) {
+            carRepository.deleteById(id);
+        }
     }
 }
